+++ cmem_plugin_base/dataintegration/description.py
@@ -1,7 +1,10 @@
 """Classes for describing plugins"""
 import inspect
+from base64 import b64encode
 from dataclasses import dataclass, field
 from inspect import _empty
+from mimetypes import guess_type
+from pkgutil import get_data
 from typing import Optional, List, Type, Any
 
 from cmem_plugin_base.dataintegration.plugins import WorkflowPlugin, TransformPlugin
@@ -11,6 +14,53 @@ from cmem_plugin_base.dataintegration.types import (
     PluginContextParameterType,
 )
 from cmem_plugin_base.dataintegration.utils import generate_id
+
+
+class Icon:
+    """An Icon.
+
+    :param file_name: The name of the icon file, e.g. 'icon.svg', should be in the
+        form of a relative filename, using '/' as the path separator. The parent
+        directory name '..' is not allowed, and nor is a rooted name
+        (starting with a '/').
+    :param package: (Optional) The name of the package, e.g.
+        'cmem-plugin-my.workspace', should be in standard module format. For a local
+        file from the same module, you can use package=__package__.
+    """
+
+    def __init__(
+        self,
+        file_name: str,
+        package: str
+    ) -> None:
+        self.file_name = file_name
+
+        self.package = package
+
+        try:
+            self.data = get_data(self.package, file_name)
+        except FileNotFoundError as error:
+            raise FileNotFoundError(
+                f"No icon file '{self.file_name}' in package {self.package} found."
+            ) from error
+        if self.data is None:
+            raise FileNotFoundError(
+                f"No icon file '{self.file_name}' in package {self.package} found."
+            )
+
+        self.mime_type = guess_type(self.file_name)[0]
+        if self.mime_type is None:
+            raise ValueError(
+                f"Could not guess the mime type of the file '{self.file_name}'."
+            )
+        if not self.mime_type.startswith("image/"):
+            raise ValueError(
+                f"Guessed mime type '{self.mime_type}' does not start with 'image/'."
+            )
+
+    def __str__(self):
+        data_base64 = b64encode(self.data).decode()
+        return f"""data:{self.mime_type};base64,{data_base64}"""
 
 
 class PluginParameter:
@@ -57,11 +107,12 @@ class PluginDescription:
     :param documentation: Documentation for this plugin in Markdown.
     :param categories: The categories to which this plugin belongs to.
     :param parameters: Available plugin parameters
+    :param icon: An optional custom plugin icon.
     """
 
     # pylint: disable=too-many-instance-attributes
 
-    def __init__(
+    def __init__(  # pylint: disable=too-many-arguments
         self,
         plugin_class,
         label: str,
@@ -70,6 +121,7 @@ class PluginDescription:
         documentation: str = "",
         categories: Optional[List[str]] = None,
         parameters: Optional[List[PluginParameter]] = None,
+        icon: Optional[Icon] = None
     ) -> None:
         #  Set the type of the plugin. Same as the class name of the plugin
         #  base class, e.g., 'WorkflowPlugin'.
@@ -103,6 +155,7 @@ class PluginDescription:
             self.parameters = []
         else:
             self.parameters = parameters
+        self.icon = icon
 
 
 @dataclass
@@ -174,7 +227,8 @@ class Plugin:
         do not need to add a first level heading to the markdown since the
         documentation rendering component will add a heading anyway.
     :param categories: The categories to which this plugin belongs to.
-    :param parameters: Available plugin parameters
+    :param parameters: Available plugin parameters.
+    :param icon: Optional custom plugin icon.
     """
 
     plugins: list[PluginDescription] = []
@@ -187,11 +241,13 @@ class Plugin:
         documentation: str = "",
         categories: Optional[List[str]] = None,
         parameters: Optional[List[PluginParameter]] = None,
+        icon: Optional[Icon] = None
     ):
         self.label = label
         self.description = description
         self.documentation = documentation
         self.plugin_id = plugin_id
+        self.icon = icon
         if categories is None:
             self.categories = []
         else:
@@ -210,6 +266,7 @@ class Plugin:
             documentation=self.documentation,
             categories=self.categories,
             parameters=self.retrieve_parameters(func),
+            icon=self.icon
         )
         Plugin.plugins.append(plugin_desc)
         return func